Refactor dan Optimasi Code

+++ database/migrations/2026_07_23_062946_add_is_late_submit_to_reports_table.php
@@ -9,15 +9,19 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('reports', function (Blueprint $table) {
-            // Tambahkan kolom boolean untuk status telat submit akhir shift
-            $table->boolean('is_late_submit')->default(false)->after('is_late');
+
+            if (!Schema::hasColumn('reports', 'is_late_submit')) {
+                $table->boolean('is_late_submit')->default(false);
+            }
         });
     }
 
     public function down(): void
     {
         Schema::table('reports', function (Blueprint $table) {
-            $table->dropColumn('is_late_submit');
+            if (Schema::hasColumn('reports', 'is_late_submit')) {
+                $table->dropColumn('is_late_submit');
+            }
         });
     }
 };
